pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "network-bias-dynamics"
version = "0.1.0"
description = "Simulation suite for studying bias propagation in networked averaging dynamics"
authors = [{name = "OpenAI Codex"}]
license = {text = "MIT License"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "numpy",
    "matplotlib",
    "pandas",
    "pyyaml",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
    "black",
]

[project.urls]
Homepage = "https://example.com/network-bias-dynamics"

[tool.setuptools.packages.find]
where = ["src"]
